Data validation Completed

+++ networksecurity/utils/main_utils/utils.py
@@ -14,15 +14,14 @@ def read_yaml_file(file_path: str) -> dict:
     except Exception as e:
         raise NetworkSecurityException(e,sys)
     
-
-def write_yaml_file(file_path:str,content:object, replace:bool=False)-> None:
-    try:
-        if replace:
-            if os.path.exists(file_path):
-                os.remove(file_path)
-            os.makedirs(os.path.dirname(file_path), exist_ok=True)
-            with open (file_path, "w") as file:
-                yaml.dump(content,file)
     
+def write_yaml_file(file_path: str, content: object, replace: bool = False) -> None:
+    try:
+        if replace and os.path.exists(file_path):
+            os.remove(file_path)
+            
+        os.makedirs(os.path.dirname(file_path), exist_ok=True)
+        with open(file_path, "w") as file:
+            yaml.dump(content, file)
     except Exception as e:
-        raise NetworkSecurityException(e,sys)
+        raise NetworkSecurityException(e, sys)
